Create main
Starts by simply displaying our band's name (Luna Blue) until the space bar is pressed. The program will then take a list of names in a text file, scans through each name to identify each separate name then 'spins' through the names on the screen, selecting a random one thirty times before selecting the 'winner'.

I'm particularly proud of this project, not for it's complexity, but for the fact that it represents the first time I actually applied my programming knowledge to the real world. We needed to select a competition winner, and I realised I could make it more fun! I can also now simply change the text file and background image in order to reuse this project for future competitions! 

Not only that, but I put this together in under 30 minutes!